+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+## [1.10.6] - 2020-02-28
+
 ### Changed
 
 - Core: supplier name max length to 128 from 64